requirements.txt
aiogram
httpx
beautifulsoup4
apscheduler
python-dotenv
aiohttp
